closeByEsc progress
