+++ src/components/Stack/Stack.tsx
@@ -5,7 +5,7 @@ import largeCurrencyIcon from "../../images/Subtract.svg";
 import { Button } from "@ya.praktikum/react-developer-burger-ui-components";
 import OrderDetails from "../OrderDetails/OrderDetails";
 import Popup from "../Popup/Popup";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 
 function Stack(props: any) {
   const data = Array.from(Object.values(props));
@@ -23,6 +23,17 @@ function Stack(props: any) {
   const toggleModal = () => {
     setModal(!modal);
   };
+
+  useEffect(() => {
+    function handleEscapeKey(event: KeyboardEvent) {
+      if (event.code === "Escape") {
+        toggleModal();
+      }
+    }
+
+    document.addEventListener("keydown", handleEscapeKey);
+    return () => document.removeEventListener("keydown", handleEscapeKey);
+  });
 
   return (
     <>
